Use intptr_t instead of int for pointer storage

+++ src/arch/Threads/Threads_Pthreads.cpp
@@ -39,14 +39,14 @@ uint64_t ThreadImpl_Pthreads::GetThreadId() const
 	return threadHandle;
 }
 
-int ThreadImpl_Pthreads::Wait()
+intptr_t ThreadImpl_Pthreads::Wait()
 {
 	void *val;
 	int ret = pthread_join( thread, &val );
 	if( ret )
 		RageException::Throw( "pthread_join: %s", strerror(errno) );
 
-	return (int) val;
+	return (intptr_t)val;
 }
 
 ThreadImpl *MakeThisThread()
